Dispaly a table on the committee indexp age
